feat: implement close treatment functionality with user input and error handling improvements
